requirements.txt
garminconnect>=0.2.8
python-dotenv>=1.0.0
